corrections for last manual merge.

+++ lib/jxl/modular/encoding/context_predict.h
@@ -331,7 +331,8 @@ class MATreeLookup {
     }                                                                          \
     bool p0 = properties[node.property0] <= node.splitval0;                    \
     uint32_t off0 = properties[node.properties[0]] <= node.splitvals[0];       \
-    uint32_t off1 = 2 | (properties[node.properties[1]] <= node.splitvals[1]); \
+		/* fix warning C4806: '|': unsafe operation: no value of type 'bool' promoted to type 'int' can equal the given constant */ \
+    uint32_t off1 = 2 | uint32_t(properties[node.properties[1]] <= node.splitvals[1]); \
     pos = node.childID + (p0 ? off1 : off0);                                   \
   }
